+++ src/main/java/com/example/back_health_monitor/user/UserResponseDTO.java
@@ -2,7 +2,7 @@ package com.example.back_health_monitor.user;
 
 import java.time.LocalDate;
 
-public record UserResponseDTO(String cpf, String patient_name, LocalDate birth_date, Integer emergency_contact, String password, String address) {
+public record UserResponseDTO(String cpf, String patient_name, LocalDate birth_date, Long emergency_contact, String password, String address) {
 
     public UserResponseDTO(User patient){
         this(
